ZA: Move the search view in views.py to its own file
