+++ pombola/south_africa/views/__init__.py
@@ -3,4 +3,5 @@ from .geolocalization import *
 from .home import *
 from .organisations import *
 from .person import *
+from .search import *
 from .views import *
