+++ pombola/south_africa/views/views.py
@@ -21,7 +21,7 @@ from django.utils.translation import ugettext_lazy as _
 from django.db.models import Q
 from django.contrib.contenttypes.models import ContentType
 
-from haystack.query import RelatedSearchQuerySet, SQ
+from haystack.query import RelatedSearchQuerySet
 from haystack.inputs import AutoQuery
 
 from info.views import InfoBlogView
@@ -36,47 +36,10 @@ from slug_helpers.views import SlugRedirect
 from pombola.core import models
 from pombola.core.views import (
     CommentArchiveMixin, PersonSpeakerMappingsMixin)
-from pombola.search.views import SearchBaseView
 
 from pombola.interests_register.models import Release, Category, Entry
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 from django_date_extensions.fields import ApproximateDate
-
-
-class SASearchView(SearchBaseView):
-
-    def __init__(self, *args, **kwargs):
-        super(SASearchView, self).__init__(*args, **kwargs)
-        del self.search_sections['speeches']
-        self.search_sections['questions'] = {
-            'model': Speech,
-            'title': 'Questions and Answers',
-            'filter': {
-                'args': [SQ(tags='question') | SQ(tags='answer')],
-            }
-        }
-        self.search_sections['committee'] = {
-            'model': Speech,
-            'title': 'Committee Appearance',
-            'filter': {
-                'kwargs': {
-                    'tags': 'committee'
-                }
-            }
-        }
-        self.search_sections['hansard'] = {
-            'model': Speech,
-            'title': 'Hansard',
-            'filter': {
-                'kwargs': {
-                    'tags': 'hansard'
-                }
-            }
-        }
-        self.section_ordering.remove('speeches')
-        self.section_ordering += [
-            'questions', 'committee', 'hansard'
-        ]
 
 
 class SANewsletterPage(InfoPageView):
